flask-app/requirements.txt
flask
gunicorn
sqlalchemy
psycopg2-binary
requests
python-dotenv
dotenv
werkzeug
